Remove import codegen from parser submodule
